// web/splive/src/app/hpft/app/models/api/patient-data-models.ts
export class PatientFilterRequest {
    fname: string;
    lname: string;
    mobno: string;
    spid: number;
    bedno: string;
    status: number;
    patientregno: string;
    admittedon: Date;
}


//Patient List Response for display first list
export class PatientDetaListResponse {
    patientid: number;
    admissionid: number;
    cpmid: number;
    fname: string;
    lname: string;
    mobno: string;
    bedno: string;
    status: number;
    spid: number;
    patientregno: string;
    drincharge: number;
    admittedon: Date;
    dischargedon: Date;
}
// ...
export class AdmissionStatusRequest {
    admissionid: number;
    status: number;
    dischargedon: Date;
}